pyproject.toml
[project]
name = "web2llm"
version = "0.2.0"
description = "A tool to scrape web content into clean Markdown for LLMs."
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
keywords = ["scraper", "llm", "markdown", "web scraping", "pdf", "github"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Text Processing",
    "Topic :: Utilities",
    "Environment :: Console",
]

dependencies = [
    "requests",
    "beautifulsoup4",
    "lxml",
    "markdownify",
    "GitPython",
    "pdfplumber",
    "PyYAML",
    "pathspec",
]

[project.urls]
Homepage = "https://github.com/herruzo99/web2llm"
"Bug Tracker" = "https://github.com/herruzo99/web2llm/issues"

[project.scripts]
web2llm = "web2llm.cli:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
]
dev = [
    "ruff",
    "pre-commit",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["web2llm*"]
exclude = ["tests*"]

[tool.setuptools]
package-data = {"web2llm" = ["default_config.yaml"]}


[tool.ruff]
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
pythonpath = ["."]
